requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
pymysql


pip install Flask-JWT-Extended
